pinctrl: qcom: Add MSM7X30 pin controller
MSM7X30 uses version 1 of the MSM GPIO chip, where the register mapping and
use is different. It also uses ProcComm to handle pin requests and mux
functions.

This is initially imported from the LA.AF.1.1_rb1.17, however it is a complete
makeover to support device tree, dynamic IO mapping and pin control.

Change-Id: I2508222dcb9265ead0fb710f8adc88d70906f517