clk: qcom: Fix RCG calculations when root is not enabled
When root is not enabled, the calculation is still returned as the registers
written, but this causes clock enable issues as the root enable bit is only set
in set_rate. This is caused because Linux clock framework caches the rate and if
the new rate is same as old rate, it will skip set_rate. To workaround this,
return clock rate as 0 so drivers that do use set_rate can update the clock
frequency and enable root bit.

Change-Id: Ic1487155b7dcec3fff4c0008105bc397be6c064b